Initial Commit
This product is not FREE. The source code provided here is only for educational means. To use this product you still need a valid license purchased from Wusik.com
Also, to re-use this code you need a commercial license from Wusik.com, contact support for details.
Do not share binaries of this code in any way, unless you have a license from Wusik.com that fits this need.
Do not distribute this code in any way, but feel free to learn from it.
In order to compile this code you need to download JUCE and understand how to use it.
https://github.com/WeAreROLI/JUCE
I have provided videos on how to do that on my Youtube Channel, starting with the video below.
https://youtu.be/yqAt3Qo1Lv8